requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
authlib>=1.3.0
python-multipart>=0.0.12
pymongo>=4.10.0
pydantic>=2.10.0
itsdangerous>=2.0.0
python-dotenv>=1.0.0
